Moves logging setup in manage.py
