+++ services/server/app/__init__.py
@@ -8,11 +8,7 @@ from flask_cors import CORS
 from flask_migrate import Migrate
 from flask_sqlalchemy import SQLAlchemy
 
-import coloredlogs
-import logging
-import logging.config
 import os
-import yaml
 
 
 admin = Admin(template_mode="bootstrap3")
@@ -22,22 +18,7 @@ bcrypt = Bcrypt()
 migrate = Migrate()
 
 
-def setup_logging():
-    with open("/usr/src/app/logging.yml", "r") as fp:
-        try:
-            config_dict = yaml.safe_load(fp.read())
-            logging.config.dictConfig(config_dict)
-            coloredlogs.install()
-        except Exception as e:
-            print(e)
-            print("There is an error in the logging configuration")
-        else:
-            logging.info("Logging setup finished successfully")
-
-
 def create_app(environemnt):
-    setup_logging()
-
     app = Flask(__name__)
     app.config.from_object(cfg_map[environemnt])
